Set iocs found in IEX calls to static and fix GCM calls
IEX calls just contain powershell, and so have the same false positive
problems that the raw powershell script itself has. IOCs found in IEX
are now static.

Sample 1954ce6974e8203214c9eb20c2d73bd765e8d5599b9a02b3656845a50f61a634
had an issue where a download command was missed within an iex call,
and after changing IEX to static the ioc in that download was no longer
tagged as dynamic. Overriding Get-Command so that both an override and
it's original Cmdlet are not listed together in it's output fixed that
issue, and now the download is captured in the actions correctly.